Add users to db and serializers as well
