+++ app/controllers/api/v1/users_controller.rb
@@ -1,0 +1,2 @@
+class Api::V1::UsersController < ApplicationController
+end
