+++ app/helpers/api/v1/users_helper.rb
@@ -1,0 +1,2 @@
+module Api::V1::UsersHelper
+end
