+++ app/models/category.rb
@@ -1,4 +1,4 @@
 class Category < ApplicationRecord
   has_many :products
-  validates :product_name , presence: true
+  validates :category_name, presence: true
 end
