+++ db/migrate/20230507134231_create_users.rb
@@ -1,0 +1,11 @@
+class CreateUsers < ActiveRecord::Migration[7.0]
+  def change
+    create_table :users do |t|
+      t.string :username
+      t.string :password_digest
+      t.string :email
+
+      t.timestamps
+    end
+  end
+end
