+++ test/controllers/api/v1/users_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class Api::V1::UsersControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
